Replace the user-management role with a playbook
The playbook at ansible/playbooks/usermgmt replaces the role "users".
It can be invoked with a single or multiple users and/or groups.

The new playbook expects that each input file describes a single user
or group, whereas the previous role expected a single yaml file
containing the complete list of users or groups.