Add workarounds for cachefilesd and autofs
Autofs sometimes reports a false success with the message
setautomntent: lookup(sss): setautomntent: Connection refused
This might be caused by starting up too quickly after sssd. As a
workaround, a 10-second startup delay is introduced.

Cachefilesd frequently fails on startup, but succeeds on a manual
attempt later. As a workaround, the service is now configured to retry
automatically.